% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanhop.R
\docType{data}
\name{cleaned_cleanhop_counts}
\alias{cleaned_cleanhop_counts}
\title{Counts table generate after the use of cleanhop function on example single cell RNAseq data reduced to 8 genes and 768 cells}
\format{a data frame with 8 row and 768 columns:
\describe{
    \item{columns}{Cell_ID}
}}
\usage{
cleaned_cleanhop_counts
}
\description{
A dataset containing single cell RNAseq gene counts for 8 genes in 768 cells barcoded with Nextera XT indexes
}
\keyword{datasets}
