% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanhop.R
\name{cleanhop}
\alias{cleanhop}
\title{Remove potential misassigned reads due to index hopping in single cell RNAseq data generated with combinatorial dual indexes.}
\usage{
cleanhop(sce, frct_value = 0.005)
}
\arguments{
\item{sce}{A SingleCellExperiement object.}

\item{frct_value}{Fraction of total reads to substract. Default value = 0.005.}
}
\value{
\code{sce} A SingleCellExperiement object.
}
\description{
Remove potential misassigned reads due to index hopping in single cell RNAseq data generated with combinatorial dual indexes.
}
\examples{
sce <- SingleCellExperiment(assays = list(counts = as.matrix(cleanhop_counts)),
                            colData = cleanhop_annotations)
cleaned_sce = cleanhop(sce)
cleaned_sce = cleanhop(sce, frct_value = 0.01)
}
