% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanhop.R
\docType{data}
\name{cleanhop_annotations}
\alias{cleanhop_annotations}
\title{Example single cell RNAseq data reduced to 8 genes and 768 cells, column annotations}
\format{a data frame with 768 row and 4 columns:
\describe{
    \item{Run_ID}{Sequencing run associated with expression data (counts) for each cell}
    \item{Lane_ID}{Sequencing lane associated with expression data (counts) for each cell}
    \item{i5_ID}{First index associated to each cell}
    \item{i7_ID}{Second index associated to each cell}
    \item{row}{Cell_ID}
}}
\usage{
cleanhop_annotations
}
\description{
A dataset containing single cell RNAseq gene annotation for 768 cells barcoded with Nextera XT indexes
}
\keyword{datasets}
