% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanhop.R
\docType{data}
\name{cleanhop_counts}
\alias{cleanhop_counts}
\title{Example single cell RNAseq data reduced to 8 genes and 768 cells, counts}
\format{a data frame with 8 row and 768 columns:
\describe{
    \item{columns}{Cell_ID}
}}
\usage{
cleanhop_counts
}
\description{
A dataset containing single cell RNAseq gene counts for 8 genes in 768 cells barcoded with Nextera XT indexes
}
\keyword{datasets}
