% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexplot.R
\name{indexplot}
\alias{indexplot}
\title{Create visual representation of one gene expression according to indexes combinaisons.}
\usage{
indexplot(sce, goi)
}
\arguments{
\item{sce}{A SingleCellExperiement object.}

\item{goi}{Gene of interest}
}
\value{
ggplot
}
\description{
Create visual representation of one gene expression according to indexes combinaisons.
}
\examples{
sce <- SingleCellExperiment(assays = list(counts = as.matrix(cleanhop_counts)),
                            colData = cleanhop_annotations)
indexplot(sce, "B2M")
indexplot(sce, "CCR8")
}
